requirements.txt
langchain==0.0.74
python-dotenv==0.21.1
openai==0.26.4
click==8.0
Flask==2.2.2
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0
MarkupSafe==2.1.1
Werkzeug==2.2.2
pymongo
bcrypt
Flask-WTF
email-validator
flask-cors==3.0.10
transformers